app: WIP organize functions.ts
these functions likely won't stay here but need to put them somewhere temporarily
