+++ app/src/lib/Util/store.ts
@@ -1,17 +1,16 @@
 import { readable, writable, type Readable, type Writable } from 'svelte/store';
 import {
-	type Credentials,
 	type Config,
-	type Auth,
 	type Bolt,
 	type Account,
 	type SelectedPlay,
 	type GameClient,
 	type PluginMeta,
-	configDefaults,
+	defaultConfig,
 	Game,
 	Client
 } from '$lib/Util/interfaces';
+import type { Credentials } from '$lib/Services/AuthService';
 
 // readable stores. known at starup
 export const internalUrl: Readable<string> = readable('https://bolt-internal');
@@ -22,13 +21,11 @@ export const productionClientId: Readable<string> = readable(
 // writable stores
 export const bolt: Writable<Bolt> = writable();
 export const platform: Writable<string | null> = writable('');
-export const config: Writable<Config> = writable({ ...configDefaults });
+export const config: Writable<Config> = writable(defaultConfig);
 export const credentials: Writable<Map<string, Credentials>> = writable(new Map());
 export const hasBoltPlugins: Writable<boolean> = writable(false);
 export const pluginList: Writable<{ [key: string]: PluginMeta }> = writable();
 export const clientListPromise: Writable<Promise<GameClient[]>> = writable();
-export const pendingOauth: Writable<Auth | null> = writable({});
-export const pendingGameAuth: Writable<Array<Auth>> = writable([]);
 export const rs3InstalledHash: Writable<string | null> = writable('');
 export const runeLiteInstalledId: Writable<string | null> = writable('');
 export const hdosInstalledVersion: Writable<string | null> = writable('');
